% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_all_national_cases}
\alias{get_all_national_cases}
\title{Get a detailed dataset of  all reported case  in the Netherlands}
\usage{
get_all_national_cases()
}
\value{
Returns a dataframe of results containing detalied Covid-19 cases in the Netherlands
}
\description{
Get a detailed dataset of  all reported case  in the Netherlands
}
