% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_cumilative_cases_per_municpality}
\alias{get_cumilative_cases_per_municpality}
\title{Get Covid-19 cases data broken down by municipality}
\source{
for further information look at \url{https://github.com/J535D165/CoronaWatchNL/master/data-geo/data-national/}
}
\usage{
get_cumilative_cases_per_municpality(lang = "en")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{
A tibble with the following 6 columns
\describe{
\item{Date_of_report/Datum}{}
\item{Municipality_code/Gemeentecode}{}
\item{Municipality_name/Gemeentenaam}{}
\item{Province/Provincienaam}{}
\item{Total_reported/totaalAantal}{}
\item{Hospital_admission/ziekenhuisopnameAantal}{}
\item{Deceased/overledenAantal}{}
}
}
\description{
Get Covid-19 cases data broken down by municipality
}
\examples{
\dontrun{get_cumilative_cases_per_municpality()}
}
