% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_daily_cases_per_municpality}
\alias{get_daily_cases_per_municpality}
\title{Get daily number of cases per municipality (regions)}
\usage{
get_daily_cases_per_municpality()
}
\value{

}
\description{
Get daily number of cases per municipality (regions)
}
