% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{get_deceased_cases_by_sex_age}
\alias{get_deceased_cases_by_sex_age}
\title{Get Covid-19 deceased patients data broken down by Sex and Age}
\source{
https://github.com/J535D165/CoronaWatchNL/tree/master/data-desc
}
\usage{
get_deceased_cases_by_sex_age(lang = "nl")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{
A dataframe
\describe{
\item{Datum/Date}{Date of notification}
\item{LeeftijdGroep/Age_group}{Sex: Man, Vrouw, Niet vermeld ( Ages, ranging from 0 to 95, divided in 19 groups of 5. Two additional groups are 95+ and Niet vermeld (not specified))}
\item{Geslacht/Sex}{Sex: Man, Vrouw, Niet vermeld ( man, woman, not specified, respectively)}
\item{Aantal/Count}{Number of newly diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden**) female (vrouw), male (man), and non-specified cases on the date of notification in the last 24 hours}
\item{AantalCumulatief/Total_count}{ Number of diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden) female (vrouw), male (man), and non-specified cases on the date of notification since the start of the outbreak}
}
}
\description{
Get Covid-19 deceased patients data broken down by Sex and Age
}
\examples{
get_deceased_cases_by_sex_age(lang = 'en') \%>\%
dplyr::group_by(Age_group,Sex) \%>\%
dplyr::summarize(l = dplyr::last(Total_count))

}
