% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_economy_data}
\alias{get_economy_data}
\title{Get data for companies that have requested for financial aid from the government}
\usage{
get_economy_data()
}
\value{
A dataframe with the following 3 columns:
\describe{
\item{Bedrijfsnaam}{Name of company requesting financial support*}
\item{Vestigingsplaats}{Location of company	}
\item{Uitbetaald voorschotbedrag}{The advance payment to the company in euros}
}
}
\description{
The NOW register, published by UWV, mentions the companies that requested and received an advance on the reimbursement.
}
\examples{
sum(get_economy_data()$`UITBETAALD VOORSCHOTBEDRAG`)

}
