% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icu.R
\name{get_icu_data_long}
\alias{get_icu_data_long}
\title{Get ICU patients data in a long format}
\usage{
get_icu_data_long()
}
\value{
\describe{
\item{Datum/Date}{Date of notification}
\item{Type/Type}{Type of measurement: i.e., Totaal ingezette IC's, Toename opnamen (IC), Totaal opnamen (IC), Cumulatief opnamen (IC), Toename ontslag (ziekenhuis), Cumulatief ontslag (ziekenhuis), Toename ontslag (overleden), Cumulatief ontslag (overleden), Totaal ontslag (IC).}
\item{Aantal/Count}{Number of intensive care (IC) units  with at least one Dutch COVID-19 case on the date of notification (Totaal ingezette IC's), number of newly confirmed or suspected COVID-19 IC intakes on the date of notification (Toename opnamen (IC)), total number of IC intakes on the date of notification (Totaal opnamen (IC)), total number of IC intakes since the start of the outbreak until the date of notification (Cumulatief opnamen (IC)), new and total number of IC-cases that left the hospital alive (Toename ontslag (ziekenhuis) and Cumulatief ontslag (ziekenhuis), respectively), new and total number of IC-cases that died in hospital after IC intake (Toename ontslag (overleden) and Cumulatief ontslag (overleden), respectively), and total number of discharged IC-cases that are still in hospital (Totaal ontslag (IC))}
}
}
\description{
\itemize{
\item These can be situated in either the Netherlands or in Germany.
** Deaths outside hospitals are not included in these numbers.
*** These people are discharged from the IC but are still hospitalized. Eventually, these cases will be either added to the hospital discharges or IC deaths counts.
}
}
