% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icu.R
\name{get_icu_data_wide}
\alias{get_icu_data_wide}
\title{Get ICU patients data published by NICE (Nationale Intesive Care Evaluatie) in a wide format}
\usage{
get_icu_data_wide()
}
\value{
A dataframe with the 7 following columns:
\describe{
\item{Datum/Date}{Date of notification}
\item{IngezetteICs/Used IC units}{Number of intensive care (IC) units* with at least one Dutch COVID-19 case on the date of notification}
\item{TotaalOpnamen/Total IC intakes}{Total number of IC intakes on the date of notificationn}
\item{ToenameOpnamen/New IC intakes}{New number of IC-cases that left the hospital alive on the date of notification}
\item{CumulatiefOpnamen/Cumulative IC intakes}{Total number of IC-cases that left the hospital alive since the start of the outbreak until the date of notification}
\item{CumulatiefOntslagOverleden/Cumulative IC deaths}{Total number of IC-cases that died in hospital during/after IC intake since the start of the outbreak until the date of notification}
\item{TotaalOntslagIC/Total IC discharges}{Total number of discharged IC-cases that are still in hospital*** on the date of notification}
}
}
\description{
The intensive care datasets describe the new and cumulative number of COVID-19 intensive care unit (ICU) admissions per day.
}
\details{
\itemize{
\item These can be situated in either the Netherlands or in Germany.
** Deaths outside hospitals are not included in these numbers.
*** These people are discharged from the IC but are still hospitalized. Eventually, these cases will be either added to the hospital discharges or IC deaths counts.
}
}
\examples{
get_icu_data_wide() \%>\%
head()
}
