% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_rate.R
\name{get_mortality_rate}
\alias{get_mortality_rate}
\title{Get mortality data from CBS}
\source{
cbs.nl
}
\usage{
get_mortality_rate(lang = "en")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{
A tibble with 4 variables
\describe{
\item{Sex/Geslacht}{}
\item{Age_group/Gemeentecode}{Age of the decesded person on 31 December}
\item{Periods/Perioden}{}
\item{Periods/Overledenen_aantal}{}
}
}
\description{
Get mortality data from CBS
}
\examples{
get_mortality_rate() \%>\%
dim()
}
