% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_municipal_cases}
\alias{get_municipal_cases}
\title{Get the reported count of cases in each municipality (county-region) in Netherlands by date}
\source{
for further information look at \url{https://github.com/J535D165/CoronaWatchNL/master/data-geo/data-national/}
}
\usage{
get_municipal_cases()
}
\value{
A dataframe of 8 columns
}
\description{
Get the reported count of cases in each municipality (county-region) in Netherlands by date
}
\examples{
get_municipal_cases()
}
