% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_municipal_municipal}
\alias{get_municipal_municipal}
\title{Get the reported count of cases in each municipality (county-region) in Netherlands by date}
\usage{
get_municipal_municipal()
}
\value{

}
\description{
Get the reported count of cases in each municipality (county-region) in Netherlands by date
}
\examples{
get_municipal_municipal() \%>\%
dim()
}
