% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_national_cases}
\alias{get_national_cases}
\title{Get the reported count of cases in the Netherlands by date}
\usage{
get_national_cases()
}
\value{

}
\description{
Get the reported count of cases in the Netherlands by date
}
\examples{
get_national_cases() \%>\%
dim()
}
