% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_nl_age}
\alias{get_nl_age}
\title{Get Covid-19 cases data broken down by َAge}
\usage{
get_nl_age(lang = "nl")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{
\describe{
\item{Datum/Date}{Date of notification}
\item{Bron/Source}{The id of the report}
\item{LeeftijdGroep/Age_group}{Sex: Man, Vrouw, Niet vermeld ( Ages, ranging from 0 to 95, divided in 19 groups of 5. Two additional groups are 95+ and Niet vermeld (not specified))}
\item{Type}{Type of measurement: Totaal, Ziekenhuisopname, Overleden ( total, hospitilzed, deceased, respectively)}
\item{Aantal/Count}{Number of newly diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden**) female (vrouw), male (man), and non-specified cases on the date of notification in the last 24 hours}
\item{AantalCumulatief/Total_count}{ Number of diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden) female (vrouw), male (man), and non-specified cases on the date of notification since the start of the outbreak}
}
}
\description{
Get Covid-19 cases data broken down by َAge
}
