% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_nl_sex}
\alias{get_nl_sex}
\title{Get Covid-19 cases data broken down by Sex}
\source{
https://github.com/J535D165/CoronaWatchNL/tree/master/data-desc
}
\usage{
get_nl_sex(lang = "nl")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{
\describe{
\item{Datum/Date}{Date of notification}
\item{Bron/Source}{The id of the report}
\item{Geslacht/Sex}{Sex: Man, Vrouw, Niet vermeld ( man, woman, not specified, respectively)}
\item{Type}{Type of measurement: Totaal, Ziekenhuisopname, Overleden ( total, hospitilzed, deceased, respectively)}
\item{Aantal/Count}{Number of newly diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden**) female (vrouw), male (man), and non-specified cases on the date of notification in the last 24 hours}
\item{AantalCumulatief/Total_count}{ Number of diagnosed (Totaal), hospitalized (Ziekenhuisopname), and deceased (Overleden) female (vrouw), male (man), and non-specified cases on the date of notification since the start of the outbreak}
}
}
\description{
Get Covid-19 cases data broken down by Sex
}
