% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_populatuon_per_region}
\alias{get_populatuon_per_region}
\title{Get population data of regions in the Netherlands}
\usage{
get_populatuon_per_region()
}
\value{
A dataframe with the following 7 columns
}
\description{
Get population data of regions in the Netherlands
}
\examples{
\dontrun{get_populatuon_per_region()}
}
