% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_provincial_cases}
\alias{get_provincial_cases}
\title{Get the reported count of cases in each province in Netherlands by date}
\usage{
get_provincial_cases()
}
\value{

}
\description{
Get the reported count of cases in each province in Netherlands by date
}
\examples{
get_provincial_cases() \%>\%
dim()
}
