% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_raw_national_cases}
\alias{get_raw_national_cases}
\title{Title}
\usage{
get_raw_national_cases(lang = "nl")

get_raw_national_cases(lang = "nl")
}
\arguments{
\item{lang}{Specify the language of the resulting dataframe (\code{nl} for Dutch \code{en} for English)}
}
\value{



}
\description{
Title

Title
}
