% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_reproduction_rate}
\alias{get_reproduction_rate}
\title{Get daily estimated reproduction number R}
\source{
https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/ed0699d1-c9d5-4436-8517-27eb993eab6e?tab=relations
}
\usage{
get_reproduction_rate()
}
\value{
\describe{
\item{Date}{date for which the reproduction number has been estimated}
\item{Rt_low}{lower limit 95\% confidence interval	}
\item{Rt_avg}{estimated reproduction number}
\item{Rt_avg}{upper limit 95\% confidence interval}
}
}
\description{
The reproduction number R gives the average number of people infected by one person with COVID-19.
}
\examples{
mean(reproduction_rate$Rt_avg,na.rm = T)
}
