% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_sewage_data}
\alias{get_sewage_data}
\title{Get the ratio of virus particles measured by RIVM in the sewage}
\source{
https://data.rivm.nl/geonetwork/srv/dut/catalog.search#/metadata/a2960b68-9d3f-4dc3-9485-600570cd52b9
}
\usage{
get_sewage_data()
}
\value{
A dataframe with  11 columns
}
\description{
a sample of the sewage is taken for 24 hours per treatment plant and per unit of time. These samples are analysed by RIVM researchers for the number of virus particles.
}
\examples{
mean(sewage_data$RNA_per_ml,na.rm = T)
}
