% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_suspected_patients}
\alias{get_suspected_patients}
\title{Get the number of suspected patients in the Netherlands}
\usage{
get_suspected_patients()
}
\value{
A dataframe with the following 3 columns
}
\description{
Get the number of suspected patients in the Netherlands
}
\examples{
get_suspected_patients() \%>\%
dim()
}
