% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_underlying_statistics}
\alias{get_underlying_statistics}
\title{Get statistics for the underlying condition of deceased COVID-19 patients younger than 70}
\usage{
get_underlying_statistics()
}
\value{
\describe{
\item{Datum/Date}{Date of notification}
\item{Type/Type}{Type of test measurement (i.e., Totaal gemeld, Onderliggende aandoeningen en/of zwangerschap, Geen onderliggende aandoening, Niet vermeld)}
\item{AantalCumulatief/Cumulative count}{The cumulative number of deceased COVID-19 cases younger than 70 (Totaal gemeld) with (Onderliggende aandoening en/of zwangerschap) or without (Geen onderliggende aandoening) an underlying condition and/or pregnancy, and the cumulative count of cases where it was unknown whether they had an underlying condition and/or were pregnant (Niet vermeld)}
}
}
\description{
Get statistics for the underlying condition of deceased COVID-19 patients younger than 70
}
