% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{get_waze_mobility_city_level_data}
\alias{get_waze_mobility_city_level_data}
\title{Get waze mobility city level data}
\usage{
get_waze_mobility_city_level_data()
}
\value{
A data frame
}
\description{
Get waze mobility city level data
}
