% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{get_waze_mobility_country_level_data}
\alias{get_waze_mobility_country_level_data}
\title{Get waze mobility country level data}
\usage{
get_waze_mobility_country_level_data()
}
\value{
A data frame
}
\description{
Get waze mobility country level data
}
