% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_measures_data}
\alias{get_measures_data}
\title{Get data for the measures taken by the government to fight the spread of Covid-19}
\usage{
get_measures_data(lang = "nl")
}
\arguments{
\item{lang}{}
}
\value{
\describe{
\item{measureID}{Measure ID}
\item{measureDescription}{Description of the enforced measure}
\item{lastMeasureUpdate}{Date on which the measure was last updated}
\item{area}{The area in which the measure is/was enforced}
\item{status}{	Restriction status of the enforced measure (i.e., 'events' can be: Closed, Partially closed, Partially banned, Banned, or Partially cancelled)}
\item{restrictiveMeasures}{Adherence level of the measure (i.e., measure can be: Recommended or Mandatory)}
\item{startDate}{Date on which the measure was first enforced}
\item{endDate}{Date on which the measure was no longer enforced}
\item{notes}{Notes about the measure	}
}
}
\description{
Get data for the measures taken by the government to fight the spread of Covid-19
}
\examples{
dim(measures_data)
}
