% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{get_populatuon_per_region}
\alias{get_populatuon_per_region}
\title{Get population data of regions in the Netherlands}
\usage{
get_populatuon_per_region()
}
\value{

}
\description{
Get population data of regions in the Netherlands
}
\examples{
get_populatuon_per_region() \%>\%
head()
}
